MCP(Mnufacturing events)/backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
pydantic
pgvector
scikit-learn
umap-learn
requests
beautifulsoup4
tqdm
python-dotenv
sqlalchemy-utils
openai 